requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
tldextract==5.1.2
google-generativeai==0.7.2
pandas==2.2.2
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
tenacity==8.5.0
tqdm==4.66.5
